bind.sh: move hard-coded version settings into variables

+++ bind.sh
@@ -1,16 +1,23 @@
-#! /usr/bin/env bash
+#!/bin/sh
+if test -z "${PTEX_ROOT}"
+then
+  echo 1>&2 "error: PTEX_ROOT must be set in the environment"
+  exit 1
+fi
 
-# exit when any command fails
 set -e
+set -x
 
-# keep track of the last executed command
-trap 'last_command=$current_command; current_command=$BASH_COMMAND' DEBUG
-# echo an error message before exiting
-trap 'echo "\"${last_command}\" exited with code $?."' ERR
+PTEX_MAJOR_VERSION=2
+PTEX_MINOR_VERSION=4
+PTEX_PATCH_VERSION=1
+
 
 rm -rf build
 
 astgen bind -u -v 1 -o build/ast -- -I${PTEX_ROOT}/include
 asttoc build/ast -o build -p ptex -L $PTEX_ROOT/lib -l Ptex \
         --fp Ptex --tll Ptex::Ptex_dynamic \
-        -major 2 -minor 4 -patch 0
+        -major ${PTEX_MAJOR_VERSION} \
+        -minor ${PTEX_MINOR_VERSION} \
+        -patch ${PTEX_PATCH_VERSION}
